+++ dbdaora/boolean/_tests/datastore/test_integration_service_boolean_aioredis_datastore_get_many.py
@@ -119,7 +119,7 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     ]
 
     assert entities == [fake_entity.id, fake_entity2.id]
-    assert fake_service.logger.warning.call_count == 1
+    assert fake_service.logger.warning.call_count == 2
 
 
 @pytest.mark.asyncio