+++ dbdaora/hash/_tests/mongodb/test_integration_service_hash_aioredis_mongodb_get_many_composed_key.py
@@ -183,7 +183,7 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     ]
 
     assert entities == [fake_entity, fake_entity2]
-    assert fake_service.logger.warning.call_count == 1
+    assert fake_service.logger.warning.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -280,7 +280,7 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker_with_fie
     ]
 
     assert entities == [fake_entity, fake_entity2]
-    assert fake_service.logger.warning.call_count == 1
+    assert fake_service.logger.warning.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -310,10 +310,8 @@ async def test_should_get_many_from_cache_memory_and_fallback(
         'fake:other_fake:fake3'
     )
 
-    print(fake_service.cache)
     entities = [e async for e in fake_service.get_many(('other_fake', 'fake'))]
     assert entities == [fake_entity]
-    print(fake_service.cache)
 
     entities = [
         e