+++ dbdaora/hash/_tests/test_integration_service_hash_aioredis_get_many.py
@@ -79,7 +79,7 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker(
     entities = [e async for e in fake_service.get_many('fake', 'fake2')]
 
     assert entities == [fake_entity, fake_entity2]
-    assert fake_service.logger.warning.call_count == 1
+    assert fake_service.logger.warning.call_count == 2
 
 
 @pytest.mark.asyncio
@@ -181,4 +181,4 @@ async def test_should_get_many_from_fallback_after_open_circuit_breaker_with_fie
     ]
 
     assert entities == [fake_entity, fake_entity2]
-    assert fake_service.logger.warning.call_count == 1
+    assert fake_service.logger.warning.call_count == 2